azurecloud/backend/azure-functions/api-gateway/function_app.py
"""
API Gateway Azure Function App
Handles all incoming API requests and routes to appropriate services
"""
import azure.functions as func
import logging
import json
import os
from datetime import datetime

app = func.FunctionApp(http_auth_level=func.AuthLevel.FUNCTION)

# Configuration
ORCHESTRATOR_URL = os.environ.get("ORCHESTRATOR_URL", "")
RAG_PROCESSOR_URL = os.environ.get("RAG_PROCESSOR_URL", "")


@app.route(route="health", methods=["GET"])
def health_check(req: func.HttpRequest) -> func.HttpResponse:
    """Health check endpoint"""
    return func.HttpResponse(
        json.dumps({
            "status": "healthy",
            "timestamp": datetime.utcnow().isoformat(),
            "service": "api-gateway"
        }),
        mimetype="application/json",
        status_code=200
    )


@app.route(route="chat", methods=["POST"])
def chat(req: func.HttpRequest) -> func.HttpResponse:
    """
    Main chat endpoint for GenAI Copilot
    Handles user queries and returns AI-generated responses
    """
    logging.info("Chat endpoint called")

    try:
        req_body = req.get_json()
    except ValueError:
        return func.HttpResponse(
            json.dumps({"error": "Invalid JSON body"}),
            mimetype="application/json",
            status_code=400
        )

    # Validate required fields
    if "query" not in req_body:
        return func.HttpResponse(
            json.dumps({"error": "Missing required field: query"}),
            mimetype="application/json",
            status_code=400
        )

    query = req_body.get("query")
    conversation_id = req_body.get("conversation_id", "")
    user_id = req_body.get("user_id", "anonymous")

    # Log the request for audit
    logging.info(f"Chat request from user: {user_id}, conversation: {conversation_id}")

    # TODO: Forward to orchestrator service
    response = {
        "conversation_id": conversation_id or "new-conv-" + datetime.utcnow().strftime("%Y%m%d%H%M%S"),
        "query": query,
        "response": "This is a placeholder response. Connect to orchestrator for actual processing.",
        "sources": [],
        "timestamp": datetime.utcnow().isoformat()
    }

    return func.HttpResponse(
        json.dumps(response),
        mimetype="application/json",
        status_code=200
    )


@app.route(route="documents", methods=["POST"])
def upload_document(req: func.HttpRequest) -> func.HttpResponse:
    """
    Document upload endpoint
    Triggers the ingestion pipeline
    """
    logging.info("Document upload endpoint called")

    try:
        # Get file from request
        file = req.files.get("file")
        if not file:
            return func.HttpResponse(
                json.dumps({"error": "No file provided"}),
                mimetype="application/json",
                status_code=400
            )

        # Get metadata
        metadata = {
            "filename": file.filename,
            "content_type": file.content_type,
            "uploaded_at": datetime.utcnow().isoformat(),
            "user_id": req.headers.get("X-User-ID", "anonymous")
        }

        # TODO: Upload to blob storage and trigger ingestion

        return func.HttpResponse(
            json.dumps({
                "status": "uploaded",
                "document_id": f"doc-{datetime.utcnow().strftime('%Y%m%d%H%M%S')}",
                "metadata": metadata
            }),
            mimetype="application/json",
            status_code=202
        )

    except Exception as e:
        logging.error(f"Error uploading document: {str(e)}")
        return func.HttpResponse(
            json.dumps({"error": str(e)}),
            mimetype="application/json",
            status_code=500
        )


@app.route(route="documents/{document_id}", methods=["GET"])
def get_document(req: func.HttpRequest) -> func.HttpResponse:
    """Get document metadata and status"""
    document_id = req.route_params.get("document_id")

    # TODO: Fetch from Cosmos DB
    return func.HttpResponse(
        json.dumps({
            "document_id": document_id,
            "status": "processed",
            "metadata": {},
            "chunks_count": 0
        }),
        mimetype="application/json",
        status_code=200
    )


@app.route(route="search", methods=["POST"])
def search(req: func.HttpRequest) -> func.HttpResponse:
    """
    Search endpoint for document retrieval
    Uses Azure AI Search for vector + keyword hybrid search
    """
    logging.info("Search endpoint called")

    try:
        req_body = req.get_json()
    except ValueError:
        return func.HttpResponse(
            json.dumps({"error": "Invalid JSON body"}),
            mimetype="application/json",
            status_code=400
        )

    query = req_body.get("query", "")
    filters = req_body.get("filters", {})
    top_k = req_body.get("top_k", 10)

    # TODO: Forward to RAG processor
    return func.HttpResponse(
        json.dumps({
            "query": query,
            "results": [],
            "total_count": 0,
            "timestamp": datetime.utcnow().isoformat()
        }),
        mimetype="application/json",
        status_code=200
    )


@app.route(route="conversations/{conversation_id}", methods=["GET"])
def get_conversation(req: func.HttpRequest) -> func.HttpResponse:
    """Get conversation history"""
    conversation_id = req.route_params.get("conversation_id")

    # TODO: Fetch from Cosmos DB
    return func.HttpResponse(
        json.dumps({
            "conversation_id": conversation_id,
            "messages": [],
            "created_at": datetime.utcnow().isoformat()
        }),
        mimetype="application/json",
        status_code=200
    )


@app.route(route="conversations", methods=["GET"])
def list_conversations(req: func.HttpRequest) -> func.HttpResponse:
    """List user's conversations"""
    user_id = req.headers.get("X-User-ID", "anonymous")

    # TODO: Fetch from Cosmos DB
    return func.HttpResponse(
        json.dumps({
            "user_id": user_id,
            "conversations": [],
            "total_count": 0
        }),
        mimetype="application/json",
        status_code=200
    )
