azurecloud/backend/azure-functions/api-gateway/requirements.txt
# Azure Functions dependencies
azure-functions==1.17.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-search-documents==11.4.0
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
pydantic==2.5.3
